test: seperate LSP8 tests for proxy
