+++ package.json
@@ -51,7 +51,8 @@
     "test:lsp6init": "hardhat test --no-compile tests/LSP6KeyManager/LSP6KeyManagerInit.test.ts",
     "test:lsp7": "hardhat test --no-compile tests/LSP7DigitalAsset/standard/*.test.ts",
     "test:lsp7init": "hardhat test --no-compile tests/LSP7DigitalAsset/proxy/*.test.ts",
-    "test:lsp8": "hardhat test --no-compile tests/LSP8IdentifiableDigitalAsset/*.test.ts tests/LSP8IdentifiableDigitalAsset/extensions/*.test.ts",
+    "test:lsp8": "hardhat test --no-compile tests/LSP8IdentifiableDigitalAsset/standard/*.test.ts",
+    "test:lsp8init": "hardhat test --no-compile tests/LSP8IdentifiableDigitalAsset/proxy/*.test.ts",
     "test:lsp9": "hardhat test --no-compile tests/LSP9Vault/LSP9Vault.test.ts",
     "test:lsp9init": "hardhat test --no-compile tests/LSP9Vault/LSP9VaultInit.test.ts",
     "test:lsp11": "hardhat test --no-compile tests/LSP11BasicSocialRecovery/LSP11BasicSocialRecovery.test.ts",
